+++ src/stdlib/general.rs
@@ -33,7 +33,7 @@ pub(crate) fn call(vm: &mut VM, (func, args): (RantFunctionRef, Option<Vec<RantV
   Ok(())
 }
 
-pub(crate) fn concat(vm: &mut VM, mut args: VarArgs<RantValue>) -> RantStdResult {
+pub(crate) fn cat(vm: &mut VM, mut args: VarArgs<RantValue>) -> RantStdResult {
   for val in args.drain(..) {
     vm.cur_frame_mut().write_value(val);
   }